chore: add migration script
