--- app/shaolin-scrolls/page.module.css
@@ -1,16 +1,0 @@
-.container {
-  padding: 1rem;
-}
-
-.create {
-  margin-bottom: 1rem;
-}
-
-.create form {
-  display: inline-block;
-  margin-right: 1rem;
-}
-
-.create input {
-  margin-left: 0.25rem;
-}
